+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Unreleased
 
 - [#126](https://github.com/wandersoncferreira/code-review/pull/126): remove `ignore-error` and improve message on `code-review-start`
+- [#127](https://github.com/wandersoncferreira/code-review/pull/127): introduce `code-review-github-graphql-host` variable
 
 # v0.0.4