+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#127](https://github.com/wandersoncferreira/code-review/pull/127): introduce `code-review-github-graphql-host` variable
 - [#128](https://github.com/wandersoncferreira/code-review/pull/128): Gitlab support for URLs with subgroups
 - [#130](https://github.com/wandersoncferreira/code-review/pull/130): Mention user with `C-c @` in comment buffer.
+- [#131](https://github.com/wandersoncferreira/code-review/pull/131): Support Github Enterprise 3.0 by removing unsupported features in graphql schema
 
 # v0.0.4